+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.1.3 (19 December, 2019)
+
+- Fixed runtime template injection when the `runtimeChunks` optimization is used in Webpack (#26)
+
+## 0.1.2 (18 December, 2019)
+
+- Fixed caching of Webpack loader to significantly improve performance (#22)
+
+## 0.1.1 (13 December, 2019)
+
+- Fixed usage of WDS SockJS fallback (#17)
+
 ## 0.1.0 (7 December, 2019)
 
 - Initial public release